+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### v0.2.4
+
+- new component `Stepper`
+- fix input components style
+
 ### v0.2.3
 
 - new component `Popover`